+++ packages/scoreboard/lib/src/functions/gc_filter.dart
