+++ packages/scoreboard/lib/src/screens/kriti/kriti_results_page.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:provider/provider.dart';
 
-import 'package:scoreboard/src/screens/kriti/widgets/kriti_result_card.dart';
 import 'package:scoreboard/src/stores/kriti_store.dart';
 
+import '../../functions/kriti_schedule_filter.dart';
 import '../../globals/styles.dart';
 import '../../models/kriti_models/kriti_event_model.dart';
 import '../../services/api.dart';
@@ -12,8 +12,9 @@ import '../../stores/common_store.dart';
 import '../../widgets/common/err_reload.dart';
 import '../../widgets/common/shimmer.dart';
 import '../../widgets/common/top_bar.dart';
-import 'functions/kriti_schedule_filter.dart';
-import 'widgets/kriti_filter_bar.dart';
+import '../../widgets/common/kriti_filter_bar.dart';
+import '../../widgets/cards/kriti_result_card.dart';
+
 
 
 class KritiResultsPage extends StatefulWidget {
