+++ packages/scoreboard/lib/src/screens/spardha/forms/add_result_form.dart
@@ -4,7 +4,7 @@ import '../../../functions/position.dart';
 import '../../../functions/snackbar.dart';
 import '../../../functions/validator.dart';
 import '../../../globals/colors.dart';
-import '../../../models/event_model.dart';
+import '../../../models/spardha_models/spardha_event_model.dart';
 import '../../../services/api.dart';
 import '../../../stores/result_form_store.dart';
 import '../../../widgets/add_result/custom_text_field.dart';
