+++ packages/scoreboard/lib/src/widgets/cards/expanded_results_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../globals/styles.dart';
-import '../../models/event_model.dart';
+import '../../models/spardha_models/spardha_event_model.dart';
 
 // ignore: must_be_immutable
 class ExpandedResultsCard extends StatelessWidget {
