+++ packages/scoreboard/lib/src/widgets/cards/kriti_result_card.dart
