+++ packages/scoreboard/lib/src/widgets/cards/schedule/time_venue_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import '../../../globals/styles.dart';
-import '../../../models/event_model.dart';
+import '../../../models/spardha_models/spardha_event_model.dart';
 import '../../../globals/colors.dart';
 
 class TimeVenueWidget extends StatelessWidget {
