+++ packages/scoreboard/lib/src/widgets/common/kriti_filter_bar.dart
@@ -1,16 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:provider/provider.dart';
-import 'package:intl/intl.dart';
-import 'package:scoreboard/src/globals/constants.dart';
 import 'package:scoreboard/src/stores/kriti_store.dart';
 
-import '../../../globals/colors.dart';
-import '../../../globals/enums.dart';
-import '../../../globals/styles.dart';
-import '../../../stores/common_store.dart';
-import '../../../stores/spardha_store.dart';
-import '../../../stores/static_store.dart';
+import '../../globals/colors.dart';
+import '../../globals/enums.dart';
+import '../../globals/styles.dart';
+import '../../stores/common_store.dart';
+import '../../stores/static_store.dart';
 
 class KritiFilterBar extends StatefulWidget {
   const KritiFilterBar({Key? key}) : super(key: key);
